@@ -30,7 +30,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -41,7 +41,7 @@ class MyHomePage extends StatefulWidget {
   // used by the build method of the State. Fields in a Widget subclass are
   // always marked "final".
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -54,7 +54,7 @@ class _MyHomePageState extends State<MyHomePage> {
         appBar: AppBar(
           // Here we take the value from the MyHomePage object that was created by
           // the App.build method, and use it to set our appbar title.
-          title: Text(widget.title),
+          title: Text(widget.title!),
         ),
         body: ListView(
           physics: PageScrollPhysics(),
@@ -63,11 +63,17 @@ class _MyHomePageState extends State<MyHomePage> {
           children: [
             NoInternetWidget(),
             CustomNoInternetWidget(
-              imageWidget: Lottie.asset('assets/no_internet_connection_wifi.json', width: 400, fit: BoxFit.cover),
+              imageWidget: Lottie.asset(
+                  'assets/no_internet_connection_wifi.json',
+                  width: 400,
+                  fit: BoxFit.cover),
               textWidget: Text(
                 "Internet Disconnected",
                 textAlign: TextAlign.center,
-                style: TextStyle(fontSize: 22.0, color: Colors.black54, fontWeight: FontWeight.bold),
+                style: TextStyle(
+                    fontSize: 22.0,
+                    color: Colors.black54,
+                    fontWeight: FontWeight.bold),
               ),
             ),
             NoDataFoundWidget()
