@@ -1,9 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:flutter_empty_illustration/flutter_empty_illustration.dart';
-
 void main() {
-  test('adds one to input values', () {
-
-  });
+  test('adds one to input values', () {});
 }
